chore: rename translation files
